[REFACT] move into dto.response #7
